// src/types.tsx
export type Center = {
  id: number;
  name: string;
  languages: [string];
  geographicAreas: [string];
  hasStatewideService: boolean;
  specificCommunities: string;
  areasOfService: [string];
  streetAddress: string;
  city: string;
  county: string;
  zipCode: number;
  fullAddress: number;
  latitude: number;
  longitude: number;
  email: string;
  website: string;
  phoneNumber: string;
  socialMedia: string;
}